mcp-code/requirements.txt
azure-functions>=1.20.0
httpx>=0.27.0
feedparser>=6.0.11
python-dateutil>=2.9.0.post0
anyio>=4.4.0